update macros, fix sprite problrm, refactor

+++ srcs/draw_mini_map.c
@@ -2,7 +2,7 @@
 
 static void	draw_mm_buf(t_info *info, t_mini_map *mm, int color)
 {
-	while (mm->wid_cnt < mm->mm_wid / info->col_count)
+	while (mm->wid_cnt < mm->mm_wid / (int)info->col_count)
 	{
 		info->buf[mm->mm_i][mm->mm_j++] = color;
 		mm->wid_cnt++;
@@ -29,14 +29,14 @@ void	draw_mini_map(t_info *info)
 	mm.mm_i = 5;
 	mm.mm_hght = info->hght / 6;
 	mm.mm_wid = info->wid / 6;
-	while (mm.mm_i < mm.mm_hght && mm.map_i < info->row_count)
+	while (mm.mm_i < mm.mm_hght && mm.map_i < (int)info->row_count)
 	{
 		mm.hght_cnt = 0;
-		while (mm.hght_cnt < mm.mm_hght / info->row_count)
+		while (mm.hght_cnt < mm.mm_hght / (int)info->row_count)
 		{
 			mm.map_j = 0;
 			mm.mm_j = 5;
-			while (mm.mm_j < mm.mm_wid && mm.map_j < info->col_count)
+			while (mm.mm_j < mm.mm_wid && mm.map_j < (int)info->col_count)
 			{
 				mm.wid_cnt = 0;
 				draw_items(info, &mm);
